pyproject.toml
[tool.poetry]
name = "enforcersbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.2"
aiohttp = "^3.7.4"
python-dotenv = "^0.17.1"
Flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
